modify migration

+++ database/migrations/2024_05_25_044339_create_answers_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('answers', function (Blueprint $table) {
             $table->id();
-            $table->smallInteger('range');
+            $table->tinyInteger('range');
             $table->foreignId('question_id')->constrained()->onDelete('cascade');
             $table->timestamps();
         });
